CBMC: Add spec (but no proof) of polyvec_ntt
Our type+memory safety proof in CBMC does not require any specific
knowledge of what the NTT does. We do, however, need a bound on the
coefficients of the output polynomial, in order to prove type safety
(non-overflow) when the caller subsequently adds to the output.

The definition of the NTT and invNTT bounds are, likely temporarily,
moved to poly.h, to avoid a cyclic dependency between headers.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>