Move definition of KECCAK_WAY to params.h
Currently we define KECCAK_WAY in keccakf1600.h which
makes sense for us, but if one wants to use a different
Keccak backend one has to set it elsewhere.

This commits move the definition to params.h.
Then it is available even if a different Keccak backend
is used.

Context: Initially, we were considering backends with
different KECCAK_WAY, but right now every backend is
using KECCAK_WAY=4 and our mlkem code is quite
specific for that value.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>