Add support for monolithic build
This commit extends autogenerate_files.py to autogenerate a single
compilation unit importing all C source files in mlkem-native, thereby
enablig a single-CU build of mlkem-native. Such single-CU build can be
both easy to use as well as convenient for reproducible builds.

The single-CU build is exercised in `examples/monolithic_build`,
which compiles and links the single-CU version of mlkem-native with
a dummy `main`.

So far, the single-CU build is confined to the C backend. More work
is needed to integrate native backends.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>