Tests: Remove CompileOptions object
The `tests` script previously stored aspects of the command line
arguments in various places. For example, aspects pertaining
to compilation would be stored in a `CompileOptions` class, and other
aspects, such as the `auto` or `opt` settings, would be copied in the
respective test class.

This commit simplifies the `tests` script by directly working with the
command line parameters as returned by `argparse`. To access derived data,
static functions in a newly introduced helper class `Args` are used.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>