Tests: Simplify logging of failures
Previously, a test failure would call `sys.exit()`, which would
throw an exception, which would be caught and converted into a
boolean, and accumulated over all test runs.

Now that all tests are part of a single class, we can instead just
log a failure counter as a class field, and check its status once
in the end.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>