Tests: Don't drop success/failure indicators from run_scheme(s)
While we log the failures in the Tests class itself, the GH summary
printer currently still relies on the dictionary of success values
generated by run_schemes. This commit reinstates the latter. Ultimately,
a cleaner solution would be to build this dictionary as a field in
the Tests class, but this can be follow-up.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>