Monolithic build: Add functional test for both security levels
Previously, the `examples/monolithic_build/main.c` test didn't do
anything, but we merely tested that it built.

With this commit, `examples/monolithic_build/main.c` is modified to
conduct a basic key agreement test for both copies of mlkem-native
that are included.

There is still an inconvenience since we don't yet have a public
header for mlkem-native's API that would only depend on the config;
in consequence, the example has to manually define key sizes and
API declarations. This will go away in the future.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>