CBMC: Remove `STATIC_[INLINE_]TESTABLE`
STATIC_TESTABLE and STATIC_INLINE_TESTABLE were previously used as wrappers
around `static` and `static inline` that were defined as empty for a CBMC
build, thereby lifting visibility of static (inline) functions so that CBMC
would see them.

This commit achieves the same without STATIC_TESTABLE and
STATIC_INLINE_TESTABLE, both of which are removed, by defining `static`
and `INLINE` as the empty string in a CBMC build.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>