Format: Ignore symlinks
The examples in `examples/` contain many symlinks to the mlkem-native
source files. Previously, `format`, and specifically `clang-format`
invoked by it, would follow those symlinks and replace them by a copy
of the modified file if the linked file needed re-formatting. This is
not desirable: _only_ the file behind the symlink should be changed.

This commit fixes `format` by not running `clang-format` on symlinks.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>