Remove backend from default namespace
Previously, the backend type would enter the default namespace
used by mlkem-native.

This means that before the MLKEM_NAMESPACE/FIPS202_NAMESPACE macros
can be used, the backend header has to be included.

This is inconvenient for the introduction of a public API header which
solely depends on the config file, since that header would need to also
include the backend header, and hence be compiled with the same include
paths as the mlkem-native build itself. This is unintuitive and unnecessarily
complicated for the customer.

This commit removes the backend type from the default namespace. Customers
who need the backend type, or at least the target architecture, to be
reflected in the namespace, can overwrite the default namespace as desired.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>