Use uniform pattern for including config
Previously, `common.h` would explicitly include a custom config
if present, but not explicitly include the default `config.h`
otherwise. Instead, this would be left to `params.h`, which _always_
includes the default config `config.h`, but not a custom config.

This happens to be OK if custom configs and default configs use
the same header guard, but is confusing and non-uniform nonetheless.

This commit uniformly uses the pattern

```
 #if defined(MLKEM_NATIVE_CONFIG_FILE)
 #include MLKEM_NATIVE_CONFIG_FILE
 #else
 #include "config.h"
 #endif /* MLKEM_NATIVE_CONFIG_FILE */
```

to include the configuration.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>