Add marker for non-public API
This commit introduces the marker MLKEM_NATIVE_INTERNAL_API.
This marker is used in front of every declaration and implementation
of an API that has external linkage for the standard build process
of mlkem-native, but is not part of mlkem-native's public API.

Beyond being an explicit reminder to consumers that they can not
rely on the stability of an API so marked as internal, the marker
can be unfolded to the `static` keyword for monolithic builds,
thereby giving internal linkage to all internal functions. This
enables further compiler optimizations, and also removes the corresponding
symbols from the resulting monolithic object file.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>