CBMC: Reorder the preconditions in KeccakF1600_StateXORBytes contract
It seems that `__CPROVER_is_fresh()` (`memory_no_alias` in mlkem-native)
always succeeds when evaluated in the course of a function contract proof
(but not if a contract is applied). This renders the aliasing case in the
contract for KeccakF1600_StateXORBytes unused.
This commit fixes this by reorderig the clauses, which forces the first
clause to be considered in the contract proof. This can be witnessed by
breaking the aliasing case and seeing the proof failing; previously, it
wouldn't.

The constraints in the sound use of __CPROVER_is_fresh() when it appears
in a disjunction, need to be better understood. Luckily, we have no other
instances of this pattern in the library.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>