CBMC: Big-endian proofs for KeccakF1600_State{Extract,XOR}Bytes
Those functions have an explicit distinction between little and big
endian systems and therefore need a proof for either case.

Care has to be taken to allow overwriting the directives in the
host system's preprocessor; we disable the auto-detection of
endianness if SYS_LITTLE_ENDIAN or SYS_BIG_ENDIAN have been set
manually.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>