Config: Configure namespace prefix instead of namespace macro
Previously, the user could customize namespacing by providing
macros `MLKEM_NAMESPACE(sym)` and `FIPS202_NAMESPACE(sym)`.

This mode of configuration is slightly inconvenient in that
the namespace cannot be controlled from CFLAGS.

This commit addresses this, as follows: Instead of making the namespace
macro configurable, make the namespace _prefix_ configurable, and derive
the namespace macro: The config provides MLKEM_NAMESPACE_PREFIX and
FIPS202_NAMESPACE_PREfIX, and then the derived MLKEM_NAMESPACE(sym)
and FIPS202_NAMESPACE(sym) are the concatenation of that, an underscore `_`,
and the symbol name `sym`.

All configs and examples are adjusted accordingly.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>