Debug: Avoid reliance on C99 `snprintf()`
The debug functions in `debug/debug.c` so far depended on
`snprintf()`, which was added to the stdlib in C99.

To still be able to use it when C90 is used as the language
standard, the _ISOC99_SOURCE macro was manually set before
including `stdio.h`.

This is problematic for two reasons:

- If mlkem-native is used in a monolihic (single-CU) build,
  the _ISOC99_SOURCE macro may already be set, leading to an
  error.
- More problematically, the monolithic build of mlkem-native
  undefines all defines, including `_ISOC99_SOURCE`, which
  may alter or break compilation of other sources to come.

Both could in principle be fixed, but it seems better to
simply avoid reliance on `_ISOC99_SOURCE`.

This commit removes the need for `_ISOC99_SOURCE` by rewriting
`debug.c` to use `fprintf` directly, and avoiding `snprintf()`
as a helper.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>