Check that macros defines mlkem-native don't start with `_`
Most feature test macros start with an _, and we need to make sure that we
don't accidentally #undefine one of those at the end of the monobuild. This
has inadvertently happened with _ISOC99_SOURCE, which was set in debug.c
and #undef'ed at the end of the monobuild. If no mlkem-native-specific
macro starts with a _, we have an easier time catching those things.

This commit modifies `autogen` to fail on leading underscores
in macros defined in mlkem-native sources.

The only exceptions allows for the moment are the `__contract__`
and `__loop__` macros for CBMC.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>