Add basic AWS-LC integration test
This commit extends the CI with an integration test for AWS-LC.

The CI checks that mlkem-native successfully integrates into
AWS-LC by building and running basic tests of AWS-LC in FIPS
and non-FIPS mode.

We test both against a fixed commit and main. Testing against a
fixed known-good commit in AWS-LC means we note if we break the
integration. Testing against main means we note if a change in
AWS-LC breaks the integration. The latter is out of our control,
and we may need to temporarily ignore failures, but we should
be aware.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>