autogen: Autogenerate a monobuild assembly file
The `autogen` script autogenerates a single compilation unit pulling in
all *.c files in mlkem-native. This is suitable for single-CU builds of
mlkem-native _with the C backend_.

When the native backend is to be used in a monolithic build, a similar
approach is needed for assembly files.

This commit extends `autogen` to generate a single assembly file
including all assembly files in mlkem-native.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>