XKCP: Remove hardcoded unrolling config
The XKCP code offers the configuration option KeccakP1600times4_unrolling
which appears to offer control over whether the Keccakf1600 loop should be
unrolled. However, the *.c code is actually hardcoded to unrolling, hence
the config option is not functional.

This commit removes KeccakP1600times4_unrolling, and with it the entire
mlkem/fips202/native/x86_64/src/KeccakP-SIMD256-config.h.

It also inlines and removes the files
- mlkem/fips202/native/x86_64/src/KeccakP-1600-unrolling.macros
- KeccakP-1600-times4-SnP.h

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>