autogen: Add checker dangling register aliases in assembly files
The .req primitive is heavily used in mlkem-native's AArch64 assembly
for register aliases.

Since compilers usually just warn but not error out if a .req is
re-written with a different alias, it is good hygiene to .unreq all
register aliases once an assembly function has been
defined. Otherwise, there is risk of things breaking in a
hard-to-debug way when merging assembly files or having more than one
assembly function per file.

This commit extends the `autogen` script to check that register
aliases in assembly files are always undefined by the end of the
file. If they are not, it suggests a block of code to be added.
The code is not added automatically because the right placement isn't
obvious to infer.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>