autogen: Automatically namespace all labels in an assembly file
Previously, `autogen` was extended to automatically generate
namespacing #define's for all macros in assembly files.

In a similar vein, this commit extends `autogen` to also generate
namespacing `#define`s for all labels in an assembly file that are
not yet namespaced.

As with macro namespacing, this is to facilitate builds where multiple
assembly files are merged into one.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>