Avoid unused function warnings for basemul functions
The entire matrix-vector base multiplication can be replaced by
registering MLKEM_USE_NATIVE_POLYVEC_BASEMUL_ACC_MONTGOMERY_CACHED
in the backend. In this case, the functions
- poly_basemul_montgomery_cached
- basemul_cached
are unused.

In a normal build, this is not flagged because those functions have
external linkage. However, in a monobuild, they all have static linkage,
and an unused function warning is thrown when attempting a monobuild
with native backend.

Unfortunately, this is not entirely trivial to resolve because of cyclic
dependencies between the headers: On the one hand, `arith_backend.h` depends
on `poly.h` and `polyvec.h` because the `poly` and `polyvec` structures
feature in the native API. On the other hand, this commit introduces a
dependency of `poly.h` on `arith_backend.h` since we need to know if the
backend implements the basemul.

This commit introduces a `structs.h` header holding the definitions of
the `poly_*` structs. This header can be included from `poly.h`, `polyvec.h`,
and `arith_backend.h`, without introducing cyclic dependencies.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>