CBMC: Remove redundant `0 <= ...` assertions in loop invariants
Previously, we were using `int` as the type of loop indices, which
requires loop invariants of the form `0 <= i`.

With the recent switch to `unsigned` indices, those invariants are
automatic and can be removed for improved readability. This commit
does that.

Along the way, it is noticed that some of the NTT proofs still used
integral loop indices. Those are changed along the way.

However, care has to be taken to ensure that arguments to
`array_bounds(...)` are always interpreted as `int` before they
are used in bounds comparisons. For example, in the definition

```
 #define array_abs_bound(arr, lb, ub, k) \
    array_bound((arr), (lb), (ub), -(k) + 1, (k))
```

the negation `-(k)` could be unsigned if the promoted type of
`k` is `unsigned`.

To ensure that value bounds are always interpreted as `int`s,
explicit casts are inserted in the respective CBMC macros.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>