Namespacing: Suffix namespace with level for k-specific functions
Namespacing in mlkem-native is configured using MLKEM_NAMESPACE_PREFIX.
The actual symbol names were so far obtained by concatenating that prefix
with the function name used in the source code.

This commit slightly modifies the symbol name derived from
MLKEM_NAMESPACE_PREFIX, as follows: For functiona depending on
the security level, the level is infix inserted into the symbol name:
Where previously the symbol name would have been prefix_sym, it is now
prefix_{512,768,1024}_sym.

For the default namespace prefix, this qualification is redundant
yet harmless, since the default prefix already includes the level.
The use of the change is when the default prefix does _not_ include
the level, in which case the new namespacing scheme ensures common
names for k-independent functionality -- a first step towards
sharing that functionality in multi-level builds of mlkem-native.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>