Allow sharing of level-independent code in multilevel builds
Previously, there was no way to build mlkem-native once per security
level and share the large amounts of common code, notably all polynomial
arithmetic such as the NTT.

Building on the pre-work of the previous commits, this commit
adds support for sharing of level-independent code in a multilevel
build of mlkem-native:

- When the user MLKEM_NATIVE_MULTILEVEL_BUILD_WITH_SHARED, all
  level-independent code will be built, including code that may only
  be needed for another security level.
- When the user sets MLKEM_NATIVE_MULTILEVEL_BUILD_NO_SHARED, no
  level-independent code will be built.

Note that the user does not need to know which files/code is
level-independent: They can still include all files uniformly
in all types of builds. The source files are appropriately guarded
to be dropped when MLKEM_NATIVE_MULTILEVEL_BUILD_NO_SHARED is set.

The `monolithic_build_multilevel` example is adjusted to make
use of the new sharing feature.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>