Backends: Introduce one assembly file per function
This commit splits the AArch64 and x86 backend assembly files so that none
exports more than one function. This aligns mlkem-native with s2n-bignum.

Also, the commit unifies the assembly header to always emit a `.text`
and `.balign 4` directive.

Finally, the AVX2 assembly is re-arranged so that the global function
symbol is always at offset 0 from the .text section.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>