Rename polyvec -> poly_k
Originally, polyvec was exclusively home to functions dealing with
the `polyvec` structure. Recent changes have invalidated that:
In addition to functions operating on `struct polyvec`, polyvec.[ch]
hosts all poly-related functionality that depends on the security-level
(e.g. the `poly_cbd_eta[12]` wrappers).

This commit renames `polyvec.[ch]` to `poly_k.[ch]` to reflect
this change in meaning.

This change was obtained by
- Manual renaming of the files `polyvec.[ch]` -> `poly_k.[ch]`
- ```bash
  git ls-files         \
     | xargs -I {} sh  \
       -c "sed -i '' 's/polyvec\.\([ch]\)/poly_k.\1/g' {}"
  ```
- Manual updating of symlinks in
  * examples/bring_your_own_fips202
  * examples/custom_backend
- Bump CBMC OBJECT_BITS from 9 to 10 in invntt_layer,
  to counter a CBMC failure for MLKEM_K=4.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>