NTT: Simplify computation of zeta index
The n-th layer of the NTT needs the zeta table entries at indices
`2^(n-1)..2^n-1`. Previously, the initial index was computed using
a division by the `len` parameter, upholding the status of the
`layer` parameter as a pure ghost variable only needed for the
specification, but not the code itself.

This commit gives up the status of `layer` as a ghost variable
and instead uses it to compute the initial zeta index using a
shift instead of a division.

While `len` is not secret and there is, thus, no risk of
information leakage when a `div` instruction is used, it still
seems cleaner and faster to work with a shift.

Suggested-by: Rod Chapman <rodchap@amazon.co.uk>
Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>
Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>