META: Switch back to YAML instead of JSON
Originally, we used META.yml specify some mlkem-native metadata.
This was automatically processed in the `tests` script, which
introduced a dependency on pyyaml and required users to setup a
virtual environment to merely run basic tests. META.yml was thus
converted to META.json which could be parsed in Python without
additional package dependencies.

Now, `tests` no longer parses `META.json`, but instead relies on
a simple bash script `META.sh` to read the metadata; that script,
in turn, relies on basic grep and, if present, jq, for the parsing
of the META.json.

This commit reverts the metadata format back to YML and adjusts
META.sh to use yq instead of jq, if present.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>