CBMC: Add contracts to native backend
So far, the CBMC proofs covered the combination of C frontend and C
backend. While most native backend functions are drop-in replacements
of functions in the default C backend and should thus inherit their
specification, there are minor differences in signature (in particular
since #675) or functionality (e.g. the native implementation of
rej_uniform may reject certain inputs, and the frontend uses fallback
logic which is not present with the C backend) that merit explicit
contract annotations for all functions in the arithmetic and FIPS202
backend, and proofs against those specs.

This commit takes a step in this direction by adding CBMC contracts
for all functions in the native arithmetic backend, except for
`rej_uniform_native`, and proving that their call-sites still uphold
their spec when the respective MLKEM_USE_NATIVE_XXX option is set.
This is largely trivial, but it is still worth doing for
(a) documentation purposes, and since (b) the native backends operate
on raw arrays, while their call-sites operate in terms of the
`poly_xxx` structs.

The case of `rej_uniform` is more complicated and will be
handled separately.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>