FIPS202: Add shake128[x4]_init() to FIPS202 API
This commit extends the FIPS202 API by an explicit init function.
This is not used by the FIPS202 implementation shipped with mlkem-native
itself, but may be handy for consumers that wish to provide their own
FIPS202 implementation. It also makes the API more symmetric, since we
already have a `xxx_release()` API.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>