Sampling: Change fallback logic to native implementation
From a perspective of the native backend, the rejection sampling
routine `rej_uniform` is special in that it is not expected to
be replaced by native code in its entirety, but only in special cases.
Outside of those special cases, the default C implementation is used.

Previously, this fallback logic was implemented as follows:
First, the native backend would be called. Upon success, the function
would return immediately. Otherwise, it would fall back to the default
implementation. Success/Failure would be communicated through a special
return value -1.

There are two problems with this logic:
- It appears very difficult to reason about in CBMC: Specifically, when
  we call the native backend, we shift the input buffer by the amount of
  data that has already been successfully sampled, and CBMC struggles
  reasoning about that.
- We call the native backend with a potentially unaligned buffer, which
  seems unnatural. This is not an issue for the existing backends,
  because their bounds checks ensure that the native implementation only
  takes effect at the beginning of rejection sampling, where the output
  buffer is aligned; yet, alignment is not guaranteed in general.

This commit simplifies the fallback logic by invoking the native backend
only upon the first call to `rej_uniform`, when no coefficients have yet
been sampled. This solves both issues above.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>