CBMC: Add KeccakF1600[x4]_StatePermute proofs based on native code
This commit adds CBMC specifications to the FIPS202 backend API and
proves the functions depending on it -- KeccakF1600_StatePermute and
KeccakF1600x4_StatePermute -- memory-safe with respect to those
specifications.

Since the logic implementing KeccakF1600x4_StatePermute in the presence
of a FIPS202 backend depends on whether x1/x2/x4-batched Keccak is present
in the native backend, we provide 3 dummy FIPS202 backends for each case.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>