Compress: Use shift instead of division by 2-power
The scalar decompression routines so far used a division
by {16,32,1024,2048}. This is not a security concern since
even when compiled into variable-time `div` instructions
as these functions are only operating on public ciphertexts
(the one that comes out of encapsulation, not the one that comes
out of re-encryption).
Yet, it appears good style -- and it avoids having to declassify
the ciphertext in constant-time tests -- to avoid the division
altogether and use a shift instead. Note that since the operands
are unsigned, there is no difference between shift and division.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>
Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>