Set default to -march=native
In our benchmarks we always set -march=
explicitly to get the best performance on
the current platform.
While we know that this is needed, users of
mlkem-native may not and simply run the
bench script and be disappointed about
performance.

This commit sets the default -march to native.
It can be overwritten by passing a different
march flag through the CFLAGS env variable.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>