Benchmarks: Disable LTO
mlkem-native is safe to use with LTO, but consuming libraries
might still not want to set it (for example, because options
have to be uniform, and other code in the consuming library
is not safe to be used with LTO).

It is therefore better to benchmark without LTO to avoid surprise
performance drops when doing integration tests.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>