Add AVX2 native polynomial compress/decompression
This commit extends the arithmetic natice interface by the
polynomial compression routines
- poly_compress_d4
- poly_compress_d5
- poly_compress_d10
- poly_compress_d11
- poly_decompress_d4
- poly_decompress_d5
- poly_decompress_d10
- poly_decompress_d11

It also imports the native implementation in AVX2 intrinsics
provided by the official Kyber implementation.

Care has to be taken to not overflow the input buffer during
the vectorization. The official Kyber implementation _does_
overflow them, but ensures at the call-sites that sufficient
space is available, and that no relevant data is overwritten.

While this seems to work, it is difficult to reason about this
code, and it moreover prevents a unified interfaces for the native
(de)compression routines.

To remedy, this commit adds tail-loop handling for the (de)compression
routines that overflow their input buffers in the offical Kyber implementation,
reading/writing only as much as is actually needed.

A test is added which invokes the individual (de)compression routines
with minimal input buffers. When used with address sanitization, this
would catch a potential overflow.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>