Only use batched Keccak in getnoise_eta1122 if available
The previous commit has unconditionally changed
`poly_getnoise_eta1122_4x()` to use batched Keccak. Since this
generates more random data in 2/4 buffers than necessary, this
leads to a performance loss on systems where no fast batched
Keccak is available.

This commit extends the FIPS202 interface by the option
FIPS202_X4_DEFAULT_IMPLEMENTATION. This option indicates if
the batched implementation -- which is required to be implemented --
is merely a fall-back to 1-fold Keccak. In this case, the
implementation of `poly_getnoise_eta1122_4x` remains as before.

In the default FIPS202 header shipped with mlkem-native, we
unset/set FIPS202_X4_DEFAULT_IMPLEMENTATION based on the chosen
backend. Since this leads to fips202.h and the backend header
to be included at the same time, the backend needs to ensure not
to clash with the toplevel fips202.h -- this was previously
the case for the tiny_sha3 backend, which is now namespaced.

A CBMC proof is added which covers the
!FIPS202_X4_DEFAULT_IMPLEMENTATION path.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>